I added a cron class to hold all timed functions
