+++ package.json
@@ -18,6 +18,7 @@
     "get-audio-duration": "^3.1.0",
     "moment": "^2.29.4",
     "mongoose": "^6.6.5",
+    "node-cron": "^3.0.2",
     "p-queue": "^7.3.0",
     "twit": "^2.2.11",
     "web-audio-api": "^0.2.2"
